feat: add reset-progress functional

+++ components/progress.js
@@ -1,3 +1,6 @@
+import { showMessage } from './message/message.js';
+import { changeLevel } from './cards/cards.js';
+
 function saveProgress(currentLevel) {
     localStorage.setItem('currentLevel', currentLevel);
 }
@@ -6,12 +9,17 @@ function getProgress() {
     return localStorage.getItem('currentLevel');
 }
 
-function deleteProgress(){
-    localStorage.clear();
+function resetProgress(){
+    const userAnswer = confirm('Are you sure you want to reset the progress?');
+    if (userAnswer) {
+        localStorage.clear();
+        showMessage('The progress was reset. Now you can play from the very beginning😉');
+        changeLevel(1);
+    };
 }
 
 export {
     saveProgress,
     getProgress,
-    deleteProgress
+    resetProgress
 }
